% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{cayley}
\alias{cayley}
\title{Cayley's Distance}
\usage{
cayley(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Cayley's distance between \code{x} and \code{y}
}
\description{
\code{cayley} returns Cayley's distance between two full permutations.
}
\examples{
x = c(3,1,2,5,4)
y = 1:5
cayley(x,y)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
