% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{cayleyE}
\alias{cayleyE}
\title{Cayley's Distance with Ties}
\usage{
cayleyE(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Cayley's distance between the two permutations.
}
\description{
Computes Cayley's distance between two permutations, where
 items with equal ranking are now permitted. The number of items placed in the
 ith category must be the same. The number of groups must be 2 or 3, as Cayley's distance is undefined in other cases.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,2,3,3)
cayleyE(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
