% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{ham}
\alias{ham}
\title{Hamming Distance}
\usage{
ham(a, b)
}
\arguments{
\item{a, b}{integer vectors}
}
\value{
The Hamming distance between the two permutations.
}
\description{
Returns the Hamming distance between two full permutations.
}
\examples{
a = c(3,1,2,5,4)
b = 1:5
ham(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
