% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{hamE}
\alias{hamE}
\title{Hamming Distance with Ties}
\usage{
hamE(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns the Hamming distance between the two permutations.
}
\description{
Computes the Hamming distance between two permutations, where
 items with equal ranking are now permitted. The number of items placed in the
 ith category must be the same.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,2,3,3)
hamE(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
