% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{hamG}
\alias{hamG}
\title{Hamming Distance for any Number of Ties}
\usage{
hamG(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Hamming distance between the two permutations.
}
\description{
Computes Hamming distance between two permutations, where any number of items with equal rankings
 are now permitted in each ranking.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,3,4,4)
hamG(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
