% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{hamP}
\alias{hamP}
\title{Hamming Distance for Partial permutations}
\usage{
hamP(a, b, k)
}
\arguments{
\item{a, b}{integer vectors}

\item{k}{integer}
}
\value{
Returns the Hamming distance between the two partial permutations,
where only the first \code{k} items have been ranked. No ties are permitted.
}
\description{
Computes the Hamming distance between two partial permutations.
}
\examples{
a = c(3,1,2,5,4)
b = 1:5
k=3
hamP(a,b,k)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
