% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{inv}
\alias{inv}
\title{Inverse Permutation}
\usage{
inv(x)
}
\arguments{
\item{x}{an integer vector}
}
\value{
Returns the inverse permutation of a vector.
}
\description{
Computes the inverse of a permutation.
}
\examples{
a = c(3,1,2,5,4)
inv(a)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
