% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{kendE}
\alias{kendE}
\title{Kendall's Tau for Tankings with Ties}
\usage{
kendE(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Kendall's tau between the two rankings.
}
\description{
Computes Kendall's tau between two rankings, where
 items with equal ranking are now permitted. The number of items placed in the ith category must be the same.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,2,3,3)
kendE(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
