% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{kendG}
\alias{kendG}
\title{Kendall's Tau for any Number of Ties}
\usage{
kendG(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Kendall's tau between the two rankings.
}
\description{
Computes Kendall's tau between two rankings, where any number of items
with equal ranking are now permitted in each ranking.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,3,4,4)
kendG(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
