% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{kendP}
\alias{kendP}
\title{Kendall's Distance for Partial permutations}
\usage{
kendP(a, b, k)
}
\arguments{
\item{a, b}{integer vectors}

\item{k}{integer}
}
\value{
Returns Kendall's distance between two permutations x and y, where we care about the first k ranked items only
}
\description{
Computes Kendall's distance between two partial permutations.
}
\examples{
a = c(3,1,2,5,4)
b = 1:5
k=3
kendP(a,b,k)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
