% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kendR}
\alias{kendR}
\title{computes Kendall's distance}
\usage{
kendR(a, b, n)
}
\arguments{
\item{a, b}{integer vectors}

\item{n}{integer}
}
\description{
computes Kendall's distance
}
\keyword{internal}
