% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\docType{data}
\name{labour}
\alias{labour}
\title{Voting data from the 2010 UK Labour leadership election.}
\format{A data frame with 234 rows and 11 variables}
\source{
\url{https://docs.google.com/spreadsheets/d/1e-gx4Km2ywG85kJCf_byJdMZvdP4QkPHGjPKy_meO30/edit?hl=en&hl=en#gid=0}
}
\usage{
data(labour)
}
\description{
There are 5 candidates and 266 rankings, some of which are partial rankings. The candidates are Diane Abbott, Ed Balls, Andy Burnham, David Miliband and Ed Miliband. Each voter ranks at least one candidate.
}
\references{
\url{https://web.archive.org/web/20110101171158/http://www2.labour.org.uk/leadership-mps-and-meps}
}
\keyword{datasets}
