% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nij}
\alias{nij}
\title{computes the no of elements that are in the ith group in a
and the jth group in b}
\usage{
nij(a, b, i, j, n)
}
\arguments{
\item{a, b}{integer vectors}

\item{i, j, n}{integers}
}
\description{
computes the no of elements that are in the ith group in a
and the jth group in b
}
\keyword{internal}
