% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{perm}
\alias{perm}
\title{non decreasing permutations}
\usage{
perm(k, b)
}
\arguments{
\item{k, b}{integers}
}
\value{
returns all non decreasing permutations of 1:b of length k
}
\description{
generates all non decreasing permutations of 1:b of length k
}
\keyword{internal}
