% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{distance}
\alias{distance}
\title{Metrics for Rankings}
\usage{
distance(x, metric, perm = TRUE, ranktype = "full", k = 2)
}
\arguments{
\item{x}{matrix where each row is a ranking or a permutation}

\item{metric}{a distance metric, one of
\itemize{
\item \code{"kendall"}
\item \code{"ulam"}
\item \code{"spearrho"},  spearman's rho
\item \code{"spearfoot"}, spearman's footrule
\item \code{"hamming"}
\item \code{"cayley"}
}}

\item{perm}{\code{TRUE} for a matrix of permutations, \code{FALSE} for a matrix of rankings}

\item{ranktype}{indicates the type of ranking
   \itemize{
   \item \code{"full"} full ranking
   \item \code{"partial"},  partial ranking with no ties
   \item \code{"sametimes"}, ties allowed, the same number must be ranked 1st, 2nd etc
   \item \code{"anyties"}, ties allowed any number can be ranked 1st, 2nd etc
}}

\item{k}{the number of items ranked in a partial ranking}
}
\value{
Returns a matrix of the distances between all the rankings/ permutations
in the input matrix. If \code{ranktype} is \code{"anyties"} the function will work for any
type of partial ranking or ranking with ties. However using \code{"full"} for complete rankings
or \code{"partial"} for partial ranking with no ties for that kind of data will run faster.

If the ranking of some of the items is not known or given they can be left as \code{NA}.
}
\description{
Calculates the distances between a set of rankings or permutations,
using one of six metrics.
}
\examples{
x = t(matrix(replicate(10,sample(1:5,5)),ncol=10))
distance(x,metric = "spearfoot",perm = FALSE,"full")
}
\references{
\url{http://www.springer.com/gp/book/9780387962887}
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
