% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{spear}
\alias{spear}
\title{Spearman's Rho}
\usage{
spear(a, b)
}
\arguments{
\item{a, b}{integer vectors}
}
\value{
Returns Spearman's rho between the two permutations.
}
\description{
Computes Spearman's rho between two full permutations.
}
\examples{
a = c(3,1,2,5,4)
b = 1:5
spear(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
