% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{spearE}
\alias{spearE}
\title{Spearman's Rho for rankings with Ties}
\usage{
spearE(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Spearman's rho between the two rankings.
}
\description{
Computes Spearman's rho between two rankings, where
 items with equal ranking are now permitted. The number of items placed in the
 ith category must be the same.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,2,3,3)
spearE(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
