% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{spearfoot}
\alias{spearfoot}
\title{Spearman's Footrule}
\usage{
spearfoot(a, b)
}
\arguments{
\item{a, b}{integer vectors}
}
\value{
Returns Spearman's footrule between the two permutations.
}
\description{
Computes Spearman's Footrule between two full permutations.
}
\examples{
a = c(3,1,2,5,4)
b = 1:5
spearfoot(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
