% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{spearfootE}
\alias{spearfootE}
\title{Spearman's Footrule with Ties}
\usage{
spearfootE(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Spearman's footrule between the two permutations.
}
\description{
Computes Spearman's footrule between two permutations, where
 items with equal ranking are now permitted. The number of items placed in the
 ith category must be the same.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,2,3,3)
spearfootE(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
