% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{spearfootG}
\alias{spearfootG}
\title{Spearman's Footrule for any Number of Ties}
\usage{
spearfootG(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Spearman's footrule between the two permutations.
}
\description{
Computes Spearman's footrule between two permutations, where any number of items with equal permutations
are now permitted in each ranking.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,3,4,4)
spearfootG(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
