% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{spearG}
\alias{spearG}
\title{Spearman's Rho for any Number of Ties}
\usage{
spearG(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Spearman's rho between two rankings x and y
}
\description{
Computes Spearman's rho between two rankings x and y, where any number of items with equal rankings
are now permitted in each ranking. The number of items
 ranked r for the two rankings can vary.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,3,4,4)
spearG(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
