% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{ulam}
\alias{ulam}
\title{Ulam's Distance}
\usage{
ulam(a, b)
}
\arguments{
\item{a, b}{integer vectors}
}
\value{
Returns Ulam's distance between the two rankings.
}
\description{
Computes Ulam's distance between two full rankings.
}
\examples{
a = c(3,1,2,5,4)
b = c(1,2,3,4,5)
ulam(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
