% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{ulamG}
\alias{ulamG}
\title{Ulam's distance for any Number of Ties}
\usage{
ulamG(x, y)
}
\arguments{
\item{x, y}{integer vectors}
}
\value{
Returns Ulam's distance between the two permutations.
}
\description{
Computes Ulam's distance between two permutations, where any number of items with equal rankings
 are now permitted in each ranking.
}
\examples{
a = c(3,1,2,2,3)
b = c(1,2,3,4,4)
ulamG(a,b)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
