% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack2.R
\name{cayleyP}
\alias{cayleyP}
\title{Cayley's Distance for Partial permutations}
\usage{
cayleyP(a, b, k)
}
\arguments{
\item{a, b}{integer vectors}

\item{k}{integer}
}
\value{
Returns Cayley's distance between  two partial permutations,
where only the first \code{k} items have been ranked. No ties are permitted.
}
\description{
Computes Cayley's distance between two partial permutations.
}
\examples{
a = c(3,1,2,5,4)
b = 1:5
k=3
cayleyP(a,b,k)
}
\author{
Lucy Small, \email{lucy.small@ucdconnect.ie}
}
